@@ -4,40 +4,50 @@
   "description": "Codex Editor. Native JS, based on API and Open Source",
   "main": "index.js",
   "scripts": {
-    "build": "webpack --progress  --display-error-details --display-entrypoints"
+    "build": "rimraf dist && npm run build:dev",
+    "build:dev": "webpack --mode development --progress  --display-error-details --display-entrypoints"
   },
   "author": "Codex Team",
   "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/codex-team/codex.editor.git"
+  },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
+    "ancestors": "0.0.3",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.4",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-class-display-name": "^2.1.0",
     "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-env": "^1.7.0",
     "babel-runtime": "^6.26.0",
-    "css-loader": "^0.28.7",
-    "eslint": "^4.13.1",
-    "eslint-loader": "^1.9.0",
+    "css-loader": "^0.28.11",
+    "eslint": "^4.19.1",
+    "eslint-loader": "^2.0.0",
     "extract-text-webpack-plugin": "^3.0.2",
     "html-janitor": "^2.0.2",
     "path": "^0.12.7",
-    "postcss-apply": "^0.8.0",
+    "postcss-apply": "^0.10.0",
     "postcss-color-function": "^4.0.1",
     "postcss-color-hex-alpha": "^3.0.0",
-    "postcss-cssnext": "^3.0.2",
+    "postcss-cssnext": "^3.1.0",
     "postcss-custom-media": "^6.0.0",
-    "postcss-custom-properties": "^6.2.0",
+    "postcss-custom-properties": "^7.0.0",
     "postcss-custom-selectors": "^4.0.1",
-    "postcss-font-family-system-ui": "^2.1.1",
+    "postcss-font-family-system-ui": "^3.0.0",
     "postcss-font-variant": "^3.0.0",
-    "postcss-loader": "^2.0.9",
+    "postcss-loader": "^2.1.5",
     "postcss-media-minmax": "^3.0.0",
-    "postcss-nested": "^2.1.2",
-    "postcss-nested-ancestors": "^1.0.0",
-    "postcss-nesting": "^4.2.1",
+    "postcss-nested": "^3.0.0",
+    "postcss-nested-ancestors": "^2.0.0",
+    "postcss-nesting": "^6.0.0",
     "postcss-smart-import": "^0.7.6",
-    "webpack": "^3.10.0"
+    "rimraf": "^2.6.2",
+    "ts-loader": "^4.4.1",
+    "typescript": "^2.9.1",
+    "webpack": "^4.12.0",
+    "webpack-cli": "^3.0.3"
   },
   "dependencies": {}
 }
