@@ -7,29 +7,45 @@
  * | _______________________|
  */
 
+declare var Module: any;
+declare var $: any;
+
+/**
+ * DOM Elements
+ */
+interface InlineToolbarNodes {
+    wrapper?: Element; // main wrapper
+}
+
+/**
+ * CSS
+ */
+interface InlineToolbarCSS {
+  inlineToolbar: string;
+}
+
 export default class InlineToolbar extends Module {
 
-    /**
+  /**
+   * Inline Toolbar elements
+   */
+  nodes: InlineToolbarNodes = {
+      wrapper: null
+  };
+
+  /**
+   * CSS styles
+   */
+  CSS: InlineToolbarCSS = {
+      inlineToolbar: 'ce-inline-toolbar',
+  };
+
+  /**
      * @constructor
      */
     constructor({config}) {
 
         super({config});
-
-        this.nodes = {};
-
-    }
-
-    /**
-     * CSS styles
-     * @return {Object}
-     * @constructor
-     */
-    static get CSS() {
-
-        return {
-            inlineToolbar: 'ce-inline-toolbar',
-        };
 
     }
 
@@ -38,12 +54,16 @@ export default class InlineToolbar extends Module {
      */
     make() {
 
-        this.nodes.wrapper = $.make('div', InlineToolbar.CSS.inlineToolbar);
+        this.nodes.wrapper = $.make('div', this.CSS.inlineToolbar);
 
         /**
          * Append Inline Toolbar to the Editor
          */
         $.append(this.Editor.UI.nodes.wrapper, this.nodes.wrapper);
+
+    }
+
+    move() {
 
     }
 
